pyproject.toml
[tool.poetry]
name = "sensat-nlp-semantik-analiz"
version = "0.1.0"
description = ""
authors = ["Furkan Gul <frkangul@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.1.4"
pillow = "^10.2.0"
streamlit = "^1.29.0"
openai = "^1.7.0"
gradio = "^4.14.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
